# evidence/ex13_capstone/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API de Agendamento Clínico",
    "description": "Pacientes, profissionais de saúde e consultas. Dados de saúde (LGPD art. 11).",
    "version": "1.0.0"
  },
  "paths": {
...
    "/user/": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List Users",
        "operationId": "list_users_user__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserPublic"
                  },
                  "title": "Response List Users User  Get"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado"
          },
          "403": {
            "description": "Perfil/escopo sem permissão para a operação"
          },
          "429": {
            "description": "Limite de requisições excedido (veja Retry-After)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/patient/": {
      "get": {
        "tags": [
          "Patients"
        ],
        "summary": "List Patients",
        "operationId": "list_patients_patient__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientPublic"
                  },
                  "title": "Response List Patients Patient  Get"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado"
          },
          "403": {
            "description": "Perfil/escopo sem permissão para a operação"
          },
          "429": {
            "description": "Limite de requisições excedido (veja Retry-After)"
          },
          "404": {
            "description": "Recurso inexistente OU pertencente a outro usuário (indistinguíveis de propósito)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/appointment/": {
      "get": {
        "tags": [
          "Appointments"
        ],
        "summary": "Read Appointments",
        "operationId": "read_appointments_appointment__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AppointmentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppointmentPublic"
                  },
                  "title": "Response Read Appointments Appointment  Get"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado"
          },
          "403": {
            "description": "Perfil/escopo sem permissão para a operação"
          },
          "429": {
            "description": "Limite de requisições excedido (veja Retry-After)"
          },
          "404": {
            "description": "Recurso inexistente OU pertencente a outro usuário (indistinguíveis de propósito)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/availability/": {
      "get": {
        "tags": [
          "Availability (laboratório M2M)"
        ],
        "summary": "Read Availability",
        "operationId": "read_availability_availability__get",
        "security": [
          {
            "LabClientCredentials": [
              "availability:read"
            ]
          }
        ],
        "parameters": [
          {
            "name": "professional_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Professional Id"
            }
          },
          {
            "name": "day",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Data (YYYY-MM-DD), até 60 dias à frente",
              "title": "Day"
            },
            "description": "Data (YYYY-MM-DD), até 60 dias à frente"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado"
          },
          "403": {
            "description": "Perfil/escopo sem permissão para a operação"
          },
          "429": {
            "description": "Limite de requisições excedido (veja Retry-After)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppointmentPublic": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "patient_id": {
            "type": "integer",
            "title": "Patient Id"
          },
          "professional_id": {
            "type": "integer",
            "title": "Professional Id"
          },
          "date_time": {
            "type": "string",
            "format": "date-time",
            "title": "Date Time"
          },
          "status": {
            "$ref": "#/components/schemas/AppointmentStatus"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "patient_id",
          "professional_id",
          "date_time",
          "status",
          "reason"
        ],
        "title": "AppointmentPublic",
        "description": "Whitelist de campos devolvidos pela API. created_by_user_id, created_at,\nupdated_at, created_from_ip e internal_audit_note ficam de fora de propósito."
      },
      "AppointmentStatus": {
        "type": "string",
        "enum": [
          "agendada",
          "cancelada",
          "concluida"
        ],
        "title": "AppointmentStatus"
      },
...
      "AvailabilityResponse": {
        "properties": {
          "professional_id": {
            "type": "integer",
            "title": "Professional Id"
          },
          "day": {
            "type": "string",
            "format": "date",
            "title": "Day"
          },
          "slot_minutes": {
            "type": "integer",
            "title": "Slot Minutes"
          },
          "available_slots": {
            "items": {
              "type": "string",
              "format": "time"
            },
            "type": "array",
            "title": "Available Slots"
          }
        },
        "type": "object",
        "required": [
          "professional_id",
          "day",
          "slot_minutes",
          "available_slots"
        ],
        "title": "AvailabilityResponse",
        "description": "Visão do laboratório parceiro: só horários livres — nenhum dado de paciente."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PatientPublic": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "cpf": {
            "type": "string",
            "title": "Cpf"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "professional_id": {
            "type": "integer",
            "title": "Professional Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "cpf",
          "phone",
          "professional_id"
        ],
        "title": "PatientPublic"
      },
      "Role": {
        "type": "string",
        "enum": [
          "admin",
          "recepcionista",
          "profissional"
        ],
        "title": "Role"
      },
...
      "UserPublic": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "role",
          "is_active"
        ],
        "title": "UserPublic"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/user/signin"
          }
        }
      },
      "LabClientCredentials": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "scopes": {
              "availability:read": "Consultar horários livres de um profissional"
            },
            "tokenUrl": "/oauth/token"
          }
        }
      }
    }
  }
}